Fix downcoring hanging

+++ bootloader/dnc-commonlib.c
@@ -89,7 +89,7 @@ uint64_t memlimit = 0;
 static bool fastboot = 0;
 uint64_t io_limit = 0;
 uint8_t scc_att_index_range = 2;   /* 3 = 47:36, 2 = 43:32, 1 = 39:28, 0 = 35:24 */
-uint8_t downcore = 1;
+int downcore = 1;
 
 const char *node_state_name[] = { NODE_SYNC_STATES(ENUM_NAMES) };
 static struct dimm_config dimms[2]; /* 0 - MCTag, 1 - CData */
