+++ bootloader/dnc-commonlib.h
@@ -255,7 +255,7 @@ extern uint64_t memlimit;
 extern uint64_t io_limit;
 extern bool io_nonpref_high;
 extern uint8_t scc_att_index_range;
-extern uint8_t downcore;
+extern int downcore;
 
 checked const char *pr_size(uint64_t val);
 void udelay(uint32_t usecs);
